JSX Divided in Three Major Components
- Header.js
- LeftMenu.js
- EmailView.js

Next steps are to create a Component for the Search inside header
and another to be used inside the `ul` on the LeftMenu (Create a LI).

In total, there will be 8 components.

The app.css was broken into the respective components. The same shall
be done with the future components.